requirements.txt
python-telegram-bot==20.7
requests
pandas
lxml
nest-asyncio
matplotlib
python-dotenv
pytz
